open-source-project/agents/facebook_poster/requirements.txt
requests
beautifulsoup4
openai
Pillow
python-dotenv
